Feat: models, services, fachada, controllers

+++ petshare-backend/src/main/java/br/com/ufape/petshare/services/UserService.java
@@ -1,50 +1,65 @@
 package br.com.ufape.petshare.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import br.com.ufape.petshare.exceptions.UserNotFoundException;
-import br.com.ufape.petshare.model.users.User;
-import br.com.ufape.petshare.repository.UserRepository; 
+import br.com.ufape.petshare.model.User;
+import br.com.ufape.petshare.repository.UserRepository;
+import br.com.ufape.petshare.services.exceptions.ObjectNotFoundException; 
 
 @Service
-public class UserService {
-    private final UserRepository userRepository;
+public class UserService implements UserServiceInterface{
 
     @Autowired
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+    private UserRepository userRepository;
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    public User createUser(User user) {
+    @Override
+	@Transactional
+    public User saveUser(User user) {
         return userRepository.save(user);
     }
 
+    @Override
+	public User findUserById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new ObjectNotFoundException("User not found with ID = " + id));
+    }
+    
+	@Override
+    public Page<User> findUserPage(PageRequest pageRequest) {
+		return userRepository.findAll(pageRequest);
+	}
+
+    @Override
+	public List<User> getAllUsers() {
+        return userRepository.findAll();
+    }
+
+    @Override
+	@Transactional
+    public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new ObjectNotFoundException("User not found with ID = " + id);
+        }
+        userRepository.deleteById(id);
+    }
+
     public User updateUser(Long id, User updatedUser) {
+    	
         return userRepository.findById(id).map(user -> {
+        	updatedUser.getAddress().setId(user.getAddress().getId());
             user.setName(updatedUser.getName());
             user.setEmail(updatedUser.getEmail());
             user.setPhone(updatedUser.getPhone());
             user.setStatus(updatedUser.getStatus());
             user.setBornDate(updatedUser.getBornDate());
-            user.setCpf(updatedUser.getCpf());
-            user.setPassword(updatedUser.getPassword());
+            user.setAddress(updatedUser.getAddress());
             return userRepository.save(user);
-        }).orElseThrow(() -> new UserNotFoundException("Usuário não encontrado"));
-    }
-
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        }).orElseThrow(() -> new ObjectNotFoundException("User not found with ID = " + id));
     }
 }
